+++ packages/invoice-products/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'invoice-products',
-  version: '0.4.7',
+  version: '0.4.8',
   source: '@nuxt-customer-portal/invoice-products',
   dependsOn: ['core', 'invoices', 'products'],
   schema: './server/db/schema/invoice-products.ts',