+++ packages/administration/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nuxt-customer-portal/administration
 
+## 0.4.8
+
+### Patch Changes
+
+- @nuxt-customer-portal/core@0.4.8
+- @nuxt-customer-portal/ui@0.4.8
+
 ## 0.4.7
 
 ### Patch Changes