+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @nuxt-customer-portal/core
 
+## 0.4.8
+
 ## 0.4.7
 
 ## 0.4.6